Ver. 0.1.0 | Initial commit

+++ app/src/main/java/com/therxmv/dirolreader/domain/usecase/GetTdLibParametersUseCase.kt
@@ -8,7 +8,7 @@ import java.util.*
 
 class GetTdLibParametersUseCase(private val context: Context) {
     fun invoke() = TdlibParameters().apply {
-        apiId = BuildConfig.API_ID.toInt()
+        apiId = BuildConfig.API_ID
         apiHash = BuildConfig.API_HASH
         databaseDirectory = context.filesDir.absolutePath
         useMessageDatabase = true
